// MegabinFE/src/app/services/model/loginResponse.ts
/**
 * MySmartCloset API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserRoles } from './userRoles';


export interface LoginResponse { 
    token?: string | null;
    userId?: string;
    name?: string | null;
    email?: string | null;
    role?: UserRoles;
    expiresAt?: string;
}
export namespace LoginResponse {
}